fix(mobile): split AccountMenu leaves and share account fetches
Keep iPhone Account and plan chrome unchanged. Match the 20936588 split:
theme radios, workspace plans, and admin access live in one place instead
of being copied between the popover, Account screen, and sidebar.
